requirements/requirements.txt
# IP-Adapter Training Kit - Core Requirements

# Core ML/AI frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Hugging Face ecosystem
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Data handling
datasets>=2.12.0
numpy>=1.21.0
pandas>=1.5.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
omegaconf>=2.3.0

# Optional: for better performance
xformers>=0.0.20 